refactor: rename `Header`, `Footer` and `Paragraph` components to kebab-case
